+++ apps/explorer/lib/explorer/chain/cache/gas_usage.ex
@@ -5,14 +5,14 @@ defmodule Explorer.Chain.Cache.GasUsage do
 
   require Logger
 
-  @default_cache_period :timer.minutes(30)
+  @default_cache_period :timer.hours(2)
 
   use Explorer.Chain.MapCache,
     name: :gas_usage,
     key: :sum,
     key: :async_task,
     global_ttl: cache_period(),
-    ttl_check_interval: :timer.minutes(1),
+    ttl_check_interval: :timer.minutes(15),
     callback: &async_task_on_deletion(&1)
 
   alias Explorer.Chain